Refactor : ImageDiv.tsx 컴포넌트 위치 조정 반영
